modify proj1b 3 time

+++ proj1b/OffByOne.java
@@ -1,7 +1,7 @@
 public class OffByOne implements CharacterComparator {
     @Override
     public boolean equalChars(char x, char y) {
-        int diff = x - y;
+        int diff = (int) x - (int) y;
         return Math.abs(diff) == 1;
     }
 }
